fix: 모달 렌더링 코드 상위 컴포넌트로 이동 (#127)
